split nested-tabs test file
